requirements.txt
numpy
opencv-python
matplotlib
Pillow
tensorflow